requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
idna==3.4
itsdangerous==2.1.2
MarkupSafe==2.1.2
numpy==1.24.1
pandas==1.5.3
plotly==5.13.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tenacity==8.1.0
urllib3==1.26.14
Werkzeug==2.2.2
